% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/example_wfg.R
\name{example_wfg}
\alias{example_wfg}
\title{Example for creating a custum WFG test Problem.}
\description{
\strong{Step I - Specifying values for the underlying formalism}
Including scaling constants and parameter domains.

\strong{Step II - Specifying the shape functions}


\strong{Step III - Specifying transition vectors}
}

