% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bk1.R
\name{generateBK1}
\alias{BK1}
\alias{bk1}
\alias{generateBK1}
\title{BK1 test function generator.}
\usage{
generateBK1(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
BK1 test function generator.
}
\references{
wfg [45]
}

