% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/customWFG.R
\name{generateCustomWFG}
\alias{custumWFG}
\alias{custumWfg}
\alias{generateCustomWFG}
\title{Generate a custom WFG test function with WFG toolkit.}
\usage{
generateCustomWFG(x.max, S, D, A, trafos, shapeTrafos)
}
\arguments{
\item{x.max}{[\code{numeric}] \cr
Domains.}

\item{S}{[\code{numeric}] \cr
Scaling constants.}

\item{D}{[\code{numeric(1)}] \cr
Distance scaling constant.}

\item{A}{[\code{numeric}] \cr
Degeneracy constants.}

\item{trafos}{[\code{list}] \cr
List of WFG trafos.}

\item{shapeTrafos}{[\code{list}] \cr
List of WFG shape trafos.}
}
\value{
A \code{mooFunction}.
}
\description{
Generate a custom WFG test function with WFG toolkit.
}
\references{
Huband, Simon ; Hingston, Phil ; Barone, Luigi ; While, Lyndon:
A Review of Multiobjective Test Problems and a Scalable Test Problem
Toolkit. In: IEEE Trans. on Evolutionary Computation 10 (2006),
No. 5, pp. 477-506
}

