% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dgo1.R
\name{generateDGO1}
\alias{DGO1}
\alias{dgo1}
\alias{generateDGO1}
\title{DGO1 test function generator.}
\usage{
generateDGO1(in.dim = 1L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be one.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
DGO1 test function generator.
}
\references{
wfg [46]
}

