% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/dgo2.R
\name{generateDGO2}
\alias{DGO2}
\alias{dgo2}
\alias{generateDGO2}
\title{DGO2 test function generator.}
\usage{
generateDGO2(in.dim = 1L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be one.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
DGO2 test function generator.
}
\references{
wfg [46]
}

