% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fa1.R
\name{generateFA1}
\alias{FA1}
\alias{fa1}
\alias{generateFA1}
\title{FA1 test function generator.}
\usage{
generateFA1(in.dim = 3L, out.dim = 3L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be three.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be three.}
}
\value{
A \code{mooFunction}.
}
\description{
FA1 test function generator.
}
\references{
wfg [47]
}

