% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/far1.R
\name{generateFar1}
\alias{Far1}
\alias{far1}
\alias{generateFar1}
\title{Far1 test function generator.}
\usage{
generateFar1(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
Far1 test function generator.
}
\references{
wfg [48]
}

