% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fes1.R
\name{generateFES1}
\alias{FES1}
\alias{fes1}
\alias{generateFES1}
\title{FES1 test function generator.}
\usage{
generateFES1(in.dim = 30L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
FES1 test function generator.
}
\references{
wfg [49]
}

