% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fes2.R
\name{generateFES2}
\alias{FES2}
\alias{fes2}
\alias{generateFES2}
\title{FES2 test function generator.}
\usage{
generateFES2(in.dim = 30L, out.dim = 3L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be three.}
}
\value{
A \code{mooFunction}.
}
\description{
FES2 test function generator.
}
\references{
wfg [49]
}

