% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fes3.R
\name{generateFES3}
\alias{FES3}
\alias{fes3}
\alias{generateFES3}
\title{FES3 test function generator.}
\usage{
generateFES3(in.dim = 30L, out.dim = 4L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be four.}
}
\value{
A \code{mooFunction}.
}
\description{
FES3 test function generator.
}
\references{
wfg [49]
}

