% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ff1.R
\name{generateFF1}
\alias{FF1}
\alias{ff1}
\alias{generateFF1}
\title{FF1 test function generator.}
\usage{
generateFF1(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
FF1 test function generator.
}
\references{
wfg [50]
}

