% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ikk1.R
\name{generateIKK1}
\alias{IKK1}
\alias{generateIKK1}
\alias{ikk1}
\title{IKK1 test function generator.}
\usage{
generateIKK1(in.dim = 2L, out.dim = 3L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be three.}
}
\value{
A \code{mooFunction}.
}
\description{
IKK1 test function generator.
}
\references{
wfg [51]
}

