% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/im1.R
\name{generateIM1}
\alias{IM1}
\alias{generateIM1}
\alias{im1}
\title{IM1 test function generator.}
\usage{
generateIM1(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
IM1 test function generator.
}
\references{
wfg [52]
}

