% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jos1.R
\name{generateJOS1}
\alias{JOS1}
\alias{generateJOS1}
\alias{jos1}
\title{JOS1 test function generator.}
\usage{
generateJOS1(in.dim = 30L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
JOS1 test function generator.
}
\references{
wfg [53]
}

