% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/jos2.R
\name{generateJOS2}
\alias{JOS2}
\alias{generateJOS2}
\alias{jos2}
\title{JOS2 test function generator.}
\usage{
generateJOS2(in.dim = 30L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
JOS2 test function generator.
}
\references{
wfg [53]
}

