% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/le1.R
\name{generateLE1}
\alias{LE1}
\alias{generateLE1}
\alias{le1}
\title{LE1 test function generator.}
\usage{
generateLE1(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
LE1 test function generator.
}
\references{
wfg [56]
}

