% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lrs1.R
\name{generateLRS1}
\alias{LRS1}
\alias{generateLRS1}
\alias{lrs1}
\title{LRS1 test function generator.}
\usage{
generateLRS1(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
LRS1 test function generator.
}
\references{
wfg [54]
}

