% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ltdz1.R
\name{generateLTDZ1}
\alias{LTDZ1}
\alias{generateLTDZ1}
\alias{ltdz1}
\title{LTDZ1 test function generator.}
\usage{
generateLTDZ1(in.dim = 3L, out.dim = 3L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be three.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be three.}
}
\value{
A \code{mooFunction}.
}
\description{
LTDZ1 test function generator.
}
\references{
wfg [55]
}

