% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lz.R
\name{generateLZ}
\alias{generateLZ}
\title{LZ test function generator.}
\usage{
generateLZ(id, in.dim = 30L, out.dim = 2L)
}
\arguments{
\item{id}{[\code{integer(1)}] \cr
Which LZ function? Valid values are 1, 2, ..., 9}

\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space.}
}
\value{
A \code{mooFunction}.
}
\description{
LZ test function generator.
}
\references{
Li, Hui ; Zhang, Qingfu: Multiobjective Optimization Problems With
Complicated Pareto Sets, MOEA/D and NSGA-II. In: IEEE Trans.
Evolutionary Computation 13 (2009), No. 2, pp. 284-302
}

