% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mhhm1.R
\name{generateMHHM1}
\alias{MHHM1}
\alias{generateMHHM1}
\alias{mhhm1}
\title{MHHM1 test function generator.}
\usage{
generateMHHM1(in.dim = 1L, out.dim = 3L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be one.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be three.}
}
\value{
A \code{mooFunction}.
}
\description{
MHHM1 test function generator.
}
\references{
wfg [57]
}

