% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mhhm2.R
\name{generateMHHM2}
\alias{MHHM2}
\alias{generateMHHM2}
\alias{mhhm2}
\title{MHHM2 test function generator.}
\usage{
generateMHHM2(in.dim = 2L, out.dim = 3L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be three.}
}
\value{
A \code{mooFunction}.
}
\description{
MHHM2 test function generator.
}
\references{
wfg [57]
}

