% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlf1.R
\name{generateMLF1}
\alias{MLF1}
\alias{generateMLF1}
\alias{mlf1}
\title{MLF1 test function generator.}
\usage{
generateMLF1(in.dim = 1L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be one.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
MLF1 test function generator.
}
\references{
wfg [58]
}

