% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mlf2.R
\name{generateMLF2}
\alias{MLF2}
\alias{generateMLF2}
\alias{mlf2}
\title{MLF2 test function generator.}
\usage{
generateMLF2(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
MLF2 test function generator.
}
\references{
wfg [58]
}

