% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/generateMOO.R
\name{generateMOO}
\alias{MOO}
\alias{generateMOO}
\alias{moo}
\title{Genrate MOO functions}
\usage{
generateMOO(name = "", ...)
}
\arguments{
\item{name}{[\code{character(1)}] \cr
Which MOO function? (DTLZ, ZDT, LZ, UF, CF) Small letters and number.}

\item{...}{[\code{any}] \cr
Noch was schreiben}
}
\value{
A \code{mooFunction}.
}
\description{
Genrate MOO functions
}

