% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mop.R
\name{generateMOP}
\alias{MOP}
\alias{generateMOP}
\alias{generateMOP_C}
\alias{mop}
\title{MOP test function generator.}
\usage{
generateMOP(id, in.dim = 30L, out.dim = 2L)

generateMOP_C(id, in.dim = 30L, out.dim = 2L, on.infeasible = "stop")
}
\arguments{
\item{id}{[\code{integer(1)}] \cr
Which MOP function? Valid values are 1, 2, ..., 7. \cr
Which MOP_C function? Valid values are 1, 2, 3.}

\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Specified  for each MOP function.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Specified  for each MOP function.}

\item{on.infeasible}{[\code{character}] \cr
What should happen if infeasible values are evaluated? Possible values are
stop (code stops with error message) and NA (NA value is returned).}
}
\value{
A \code{mooFunction}.
}
\description{
MOP test function generator.
}
\references{
OHD. A. Van Veldhuizen, “Multiobjective Evolutionary Algorithms:
Classifications, Analyzes, and New Innovations,” Ph.D. Dissertation,
Air Force Institute of Technology, Wright-Patterson AFB, Jun. 1999.
}

