% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/qv1.R
\name{generateQV1}
\alias{QV1}
\alias{generateQV1}
\alias{qv1}
\title{QV1 test function generator.}
\usage{
generateQV1(in.dim = 30L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
QV1 test function generator.
}
\references{
wfg [59]
}

