% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sch1.R
\name{generateSch1}
\alias{SCH1}
\alias{generateSch1}
\alias{sch1}
\title{Sch1 test function generator.}
\usage{
generateSch1(in.dim = 1L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be one.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
Sch1 test function generator.
}
\references{
wfg [34], [39]
}

