% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sk1.R
\name{generateSK1}
\alias{SK1}
\alias{generateSK1}
\alias{sk1}
\title{SK1 test function generator.}
\usage{
generateSK1(in.dim = 1L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be one.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
SK1 test function generator.
}
\references{
wfg [61]
}

