% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sk2.R
\name{generateSK2}
\alias{SK2}
\alias{generateSK2}
\alias{sk2}
\title{SK2 test function generator.}
\usage{
generateSK2(in.dim = 4L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be four.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
SK2 test function generator.
}
\references{
wfg [61]
}

