% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sp1.R
\name{generateSP1}
\alias{SP1}
\alias{generateSP1}
\alias{sp1}
\title{SP1 test function generator.}
\usage{
generateSP1(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
SP1 test function generator.
}
\references{
wfg [60]
}

