% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssfyy1.R
\name{generateSSFYY1}
\alias{SSFYY1}
\alias{generateSSFYY1}
\alias{ssfyy1}
\title{SSFYY1 test function generator.}
\usage{
generateSSFYY1(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
SSFYY1 test function generator.
}
\references{
wfg [61]
}

