% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ssfyy2.R
\name{generateSSFYY2}
\alias{SSFYY2}
\alias{generateSSFYY2}
\alias{ssfyy2}
\title{SSFYY2 test function generator.}
\usage{
generateSSFYY2(in.dim = 1L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be one.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
SSFYY2 test function generator.
}
\references{
wfg [61]
}

