% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/tkly1.R
\name{generateTKLY1}
\alias{TKLY1}
\alias{generateTKLY1}
\alias{tkly1}
\title{TKLY1 test function generator.}
\usage{
generateTKLY1(in.dim = 4L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be four.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
TKLY1 test function generator.
}
\references{
wfg [62]
}

