% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/uf.R
\name{generateUF}
\alias{generateUF}
\title{UF test function generator.}
\usage{
generateUF(id, in.dim = 30L, out.dim = 2L)
}
\arguments{
\item{id}{[\code{integer(1)}] \cr
Which UF function? Valid values are 1, 2, ..., 10}

\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space.}
}
\value{
A \code{mooFunction}.
}
\description{
UF test function generator.
}
\references{
Zhang, Qingfu ; Zhou, Aimin ; Zhaoy, Shizheng ; Suganthany, Ponnuthurai N. ;
Liu, Wudong ; Tiwari, Santosh: Multiobjective optimization Test Instances
for the CEC 2009 Special Session and Competition / University of Essex and
Nanyang Technological University. 2008. - Technical Report CES-487
}

