% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vfm1.R
\name{generateVFM1}
\alias{VFM1}
\alias{generateVFM1}
\alias{vfm1}
\title{VFM1 test function generator.}
\usage{
generateVFM1(in.dim = 2L, out.dim = 3L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be three.}
}
\value{
A \code{mooFunction}.
}
\description{
VFM1 test function generator.
}
\references{
wfg [38]
}

