% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vu1.R
\name{generateVU1}
\alias{VU1}
\alias{generateVU1}
\alias{vu1}
\title{VU1 test function generator.}
\usage{
generateVU1(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
VU1 test function generator.
}
\references{
wfg [64]
}

