% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/vu2.R
\name{generateVU2}
\alias{VU2}
\alias{generateVU2}
\alias{vu2}
\title{VU2 test function generator.}
\usage{
generateVU2(in.dim = 2L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space. Must be two.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two.}
}
\value{
A \code{mooFunction}.
}
\description{
VU2 test function generator.
}
\references{
wfg [64]
}

