% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zdt.R
\name{generateZDT}
\alias{generateZDT}
\title{ZDT test function generator.}
\usage{
generateZDT(id, in.dim = 30L, out.dim = 2L)
}
\arguments{
\item{id}{[\code{integer(1)}] \cr
Which ZDT function? Valid values are 1, 2, 3, 4, 6}

\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space.}
}
\value{
A \code{mooFunction}.
}
\description{
ZDT test function generator.
}
\references{
Zitzler, Eckart ; Deb, Kalyanmoy ; Thiele, Lothar: Comparison of
Multiobjective Evolutionary Algorithms: Empirical Results. In: Evolutionary
Computation 8 (2000), pp. 173-195
}

