% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zdt1.R
\name{generateZDT1}
\alias{generateZDT1}
\title{ZDT1 test function generator.}
\usage{
generateZDT1(in.dim = 30, out.dim = 2)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space.}
}
\value{
A \code{mooFunction}.
}
\description{
ZDT1 test function generator.
}
\references{
ZDT Reference Einfuegen
}

