% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zdt4.R
\name{generateZDT4}
\alias{generateZDT4}
\title{ZDT4 test function generator.}
\usage{
generateZDT4(in.dim = 10, out.dim = 2)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space.}
}
\value{
A \code{mooFunction}.
}
\description{
ZDT4 test function generator.
}
\references{
ZDT Reference Einfuegen
}

