% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zdt6.R
\name{generateZDT6}
\alias{generateZDT6}
\title{ZDT6 test function generator.}
\usage{
generateZDT6(in.dim = 10, out.dim = 2)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space.}
}
\value{
A \code{mooFunction}.
}
\description{
ZDT6 test function generator.
}
\references{
ZDT Reference Einfuegen
}

