% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/zlt1.R
\name{generateZLT1}
\alias{ZLT1}
\alias{generateZLT1}
\alias{zlt1}
\title{ZLT1 test function generator.}
\usage{
generateZLT1(in.dim = 20L, out.dim = 2L)
}
\arguments{
\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space.}
}
\value{
A \code{mooFunction}.
}
\description{
ZLT1 test function generator.
}
\references{
wfg [11]
}

