% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/get_pareto.R
\name{getParetoSet}
\alias{getParetoFront}
\alias{getParetoSet}
\title{Wrappers to get pareto set and front from a mooFunction}
\usage{
getParetoSet(f, n)

getParetoFront(f, n)
}
\arguments{
\item{f}{[\code{function}] \cr
A \code{\link{mooFunction}}.}

\item{n}{[\code{integer(1)}] \cr.
Number of points.}
}
\description{
\code{getParetoSet} Simple wrapper for \code{attributes(f)$paretoSet(n)}, returns a \code{vector(n)}. \cr
\code{getParetoFront} Simple wrapper for calculate the pareto front, returns a \code{vector(n)}. \cr
}

