% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeWFGShapeTrafo.R
\name{makeWFGShapeTrafo}
\alias{makeWFGShapeTrafo}
\title{Make WFG Shapes}
\usage{
makeWFGShapeTrafo(arg)
}
\arguments{
\item{arg}{[\code{list}] \cr
List with \code{list(names, params)}. See examples.}
}
\value{
A [\code{list}] of WFG shapes.
}
\description{
Make WFG Shapes
}

