% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/makeWFGTrafo.R
\name{makeWFGTrafo}
\alias{makeWFGTrafo}
\title{Make WFG Transformations}
\usage{
makeWFGTrafo(arg)
}
\arguments{
\item{arg}{[\code{list}] \cr
List with \code{list(names, ids, params)}. See examples.}
}
\value{
A [\code{list}] of WFG trafos.
}
\description{
Make WFG Transformations
}

