% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plot_mooFunction.R
\name{plot.mooFunction}
\alias{plot.mooFunction}
\title{Plot a mooFunction.}
\usage{
\method{plot}{mooFunction}(x, ...)
}
\arguments{
\item{x}{[\code{function}] \cr
A \code{\link{mooFunction}}.}

\item{...}{Ignored.}

\item{resolution}{[\code{integer(1)}] \cr
Resolution for the plots. Default is 33.}

\item{plot.vars}{[\code{integer(2)}]\cr
Indices for the two variables that should be plotted. Default are the first
two variables.}

\item{const.values}{[\code{numeric}] \cr
Numeric vector with length "in.dim - 2" (one entry for each variable that is
not plotted). All variables which are not mentioned in \code{plot.vars} will
be set to a constant value given by \code{const.values}. By default the lower
bounds of the variables are used.}
}
\description{
Plot a mooFunction.
}

