% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/print_moofunction.R
\name{print.mooFunction}
\alias{print.mooFunction}
\title{Print a mooFunction.}
\usage{
\method{print}{mooFunction}(x, ...)
}
\arguments{
\item{x}{[\code{function}] \cr
A \code{\link{mooFunction}}.}

\item{...}{Ignored.}
}
\description{
Print a mooFunction.
}

