% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WFGShapes.R
\name{wfgShapeLinear}
\alias{wfgShapeConcave}
\alias{wfgShapeConvex}
\alias{wfgShapeDisconnected}
\alias{wfgShapeLinear}
\alias{wfgShapeMixed}
\title{WFG Shapes}
\usage{
wfgShapeLinear(dim, out.dim)

wfgShapeConvex(dim, out.dim)

wfgShapeConcave(dim, out.dim)

wfgShapeMixed(dim, out.dim, alpha, A)

wfgShapeDisconnected(dim, out.dim, alpha, beta, A)
}
\arguments{
\item{dim}{[\code{integer(1)}] \cr
Most shape function differ in the first (dim = 1), the last (dim = out.dim)
and the remainingout dimension. Indicates for which function the shape
function shall be returned.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space.}

\item{alpha}{[\code{numeric(1)}] \cr
Must be > 0.
The overall form of the pareto front, if alpha > 1 then it is more convex,
if alpha < 1 then more concave. When alpha = 1 the overall shape is linear.}

\item{A}{[\code{integer(1)}] \cr
Number of parts of the pareto front.}

\item{beta}{[\code{numeric(1)}] \cr
  Must be > 0.
  Where the discontinuities are. A larger value moves it to larger values of the first objective.}
}
\value{
A \code{wfgShapeFunction}.
}
\description{
wfgShapeLinear is the linear pareto front. \cr
wfgShapeConvex is the convex pareto front. \cr
wfgShapeConcave is the concave pareto front. \cr
wfgShapeMixed is a shape of the pareto front that has some convex and concave parts. \cr
wfgShapeDisconnected is a shape of the pareto front that is not continuous.\cr
}

