% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/WFGTrafos.R
\name{wfgTrafoIdentity}
\alias{wfgTrafoBFlat}
\alias{wfgTrafoBParam}
\alias{wfgTrafoBPoly}
\alias{wfgTrafoIdentity}
\alias{wfgTrafoRNonsep}
\alias{wfgTrafoRSum}
\alias{wfgTrafoSDecept}
\alias{wfgTrafoSLinear}
\alias{wfgTrafoSMulti}
\title{WFG Transformations}
\usage{
wfgTrafoIdentity(A)

wfgTrafoBPoly(alpha)

wfgTrafoBFlat(A, B, C)

wfgTrafoBParam(u, A, B, C)

wfgTrafoSLinear(A)

wfgTrafoSDecept(A, B, C)

wfgTrafoSMulti(A, B, C)

wfgTrafoRSum(w)

wfgTrafoRNonsep(A)
}
\arguments{
\item{A}{[\code{numeric(1)}] \cr
For parameter requirements see WFG Paper. Need at: \cr
wfgTrafoBFlat, wfgTrafoBParam, wfgtrafoSLinear, wfgTrafoSDecept, wfgTrafoSMulti, wfgTrafoRNonsep}

\item{alpha}{[\code{numeric(1)}] \cr
wfgTrafoBPoly: alpha > 1 biases toward 0, alpha < 1 biases toward 1.}

\item{B}{[\code{numeric(1)}] \cr
For parameter requirements see WFG Paper. Need at: \cr
wfgTrafoBFlat, wfgTrafoBParam, wfgTrafoSDecept, wfgTrafoSMulti}

\item{C}{[\code{numeric(1)}] \cr
For parameter requirements see WFG Paper. Need at: \cr
wfgTrafoBFlat, wfgTrafoBParam, wfgTrafoSDecept, wfgTrafoSMulti}

\item{u}{[\code{function}] \cr
At wfgTrafoBParamt}

\item{w}{[\code{numeric}] \cr
 At wfgTrafoRSum}
}
\value{
A \code{wfgTrafoFunction}.
}
\description{
wfgTrafoIdentity is a transformation that changes nothing.\cr
wfgTrafoBPoly is the polynomial bias transformation.\cr
wfgTrafoBFlat creates a region in search space in which all points have the same objective values.\cr
wfgTrafoBParam is the parameter-dependent transformation.\cr
wfgTrafoSLinear creates a linear shift of the true optimum.\cr
wfgTrafoSDecept creates regions in the search space that have a sub-optimal value but larger area.\cr
wfgTrafoSMulti creates many local optima.\cr
wfgTrafoRSum creates a dependence between different search-space entries.\cr
wfgTrafoRNonsep creates a dependence between objectives.\cr
}

