% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cf.R
\name{generateCF}
\alias{CF}
\alias{cf}
\alias{generateCF}
\title{CF test function generator.}
\usage{
generateCF(id, in.dim = 30L, out.dim = 2L, on.infeasible = "stop")
}
\arguments{
\item{id}{[\code{integer(1)}] \cr
Which CF function? Valid values are 1, 2, ..., 10}

\item{in.dim}{[\code{integer(1)}] \cr
Size of parameter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Size of target space. Must be two for CF 1-7 functions and three for CF 8-10 functions.}

\item{on.infeasible}{[\code{character}] \cr
What should happen if infeasible values are evaluated? Possible values are
stop (code stops with error message) and NA (NA value is returned).}
}
\value{
A \code{mooFunction}.
}
\description{
CF test function generator.
}
\examples{
#x = runif(30)
 #fun = generateCF(id = 1, in.dim = 30L, out.dim = 2L)
 #fun(x)
}
\references{
Zhang, Qingfu ; Zhou, Aimin ; Zhaoy, Shizheng ; Suganthany, Ponnuthurai N. ;
Liu, Wudong ; Tiwari, Santosh: Multiobjective optimization Test Instances
for the CEC 2009 Special Session and Competition / University of Essex and
Nanyang Technological University. 2008. - Technical Report CES-487
}

