% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mooFunction.R
\name{mooFunction}
\alias{mooFunction}
\title{Multi Objective Optimization Function}
\usage{
mooFunction(name, id, fun, in.dim, out.dim, param.set, paretoSet)
}
\arguments{
\item{name}{[\code{character(1)}] \cr
Name of function.}

\item{id}{[\code{character(1)}] \cr
Short id for the function. Must be unique to the
  function instance and should not contain any other characters than
  [a-z], [0-9] and \sQuote{-}.}

\item{fun}{[\code{function}] \cr
Function definition. Must have a single parameter x.}

\item{in.dim}{[\code{integer(1)}] \cr
Size of paramter space.}

\item{out.dim}{[\code{integer(1)}] \cr
Dimnesion of the targer space.}

\item{param.set}{[\code{\link[ParamHelpers]{ParamSet}}]\cr
Parameter set to describe (box) constraints for the function.}

\item{paretoSet}{[\code{function} | NULL] \cr
Function, that returns n points randomly distributed on the true pareto set.}
}
\value{
A \code{mooFunction} object.
}
\description{
Define a new \code{mooFunction} object.
}

